A_Level/pipeline_scripts/visual_analyzer.py
import os
import re
import json
import google.generativeai as genai
from config import settings
from . import utils

logger = utils.setup_logger(__name__)


def initialize_gemini_client():
    """Initializes the Gemini client with the API key."""
    try:
        genai.configure(api_key=settings.GOOGLE_API_KEY)
        logger.info("VisualAnalyzer: Gemini client initialized.")
        return genai.GenerativeModel(settings.GEMINI_MODEL_NAME)
    except Exception as e:
        logger.error("VisualAnalyzer: failed to initialize Gemini client — %s", e)
        return None


# The functions 'generate_visual_data_for_text' and 'request_visual_correction'
# have been removed, as this logic is now handled directly within content_creator.py.
# This file is now reserved for future, image-based analysis tasks.

def analyze_question_image(image_path):
    """
    Placeholder for future functionality to analyze a raster image of a question
    and extract features, separate from text-based generation.
    """
    logger.debug("VisualAnalyzer: analyze_question_image called with '%s'", image_path)
    raise NotImplementedError("analyze_question_image() has not been implemented yet.")
