A_Level/requirements.txt
firebase-admin
google-generativeai
numpy
pdf2image
pdfplumber
python-dotenv
requests
sympy